backend/app/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncer==0.0.8
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
boto3==1.35.76
botocore==1.35.76
bs4==0.0.2
cairocffi==1.7.1
CairoSVG==2.7.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
convcolors==2.2.0
cssselect2==0.7.0
defusedxml==0.7.1
distro==1.9.0
elevenlabs==1.13.3
extcolors==1.0.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
freetype-py==2.3.0
frozenlist==1.5.0
fsspec==2024.10.0
gradio==5.8.0
gradio_client==1.5.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.1
openai==1.55.0
openai-whisper==20240930
opencv-python-headless==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydub==0.25.1
pygal==3.0.5
Pygments==2.18.0
PyMatting==1.1.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
rembg==2.0.60
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rl-renderPM==4.0.3
rlPyCairo==0.3.0
rpds-py==0.22.3
ruff==0.8.2
s3transfer==0.10.4
safehttpx==0.1.6
scikit-image==0.24.0
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
svglib==1.5.1
sympy==1.13.1
tifffile==2024.9.20
tiktoken==0.8.0
tinycss2==1.4.0
tomlkit==0.13.2
torch==2.5.1
tqdm==4.67.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.1
yarl==1.18.3
zipp==3.21.0
